refactor: prefer `@remirror/editor-social` name
