+++ @remirror/editor-social/jest.config.js
