+++ @remirror/editor-social/package.json
@@ -1,17 +1,17 @@
 {
-  "name": "@remirror/react-editor-social",
-  "version": "0.0.0",
+  "name": "@remirror/editor-social",
+  "version": "0.13.1",
   "description": "A configurable social editor.",
-  "homepage": "https://github.com/remirror/remirror/tree/next/@remirror/react-react-editor-social",
-  "repository": "https://github.com/remirror/remirror/tree/master/@remirror/react-editor-social",
+  "homepage": "https://github.com/remirror/remirror/tree/next/@remirror/react-editor-social",
+  "repository": "https://github.com/remirror/remirror/tree/master/@remirror/editor-social",
   "license": "MIT",
   "contributors": [
     "Ifiok Jr. <ifiokotung@gmail.com>"
   ],
   "sideEffects": false,
-  "main": "dist/react-editor-social.cjs.js",
-  "module": "dist/react-editor-social.esm.js",
-  "types": "dist/react-editor-social.cjs.d.ts",
+  "main": "dist/editor-social.cjs.js",
+  "module": "dist/editor-social.esm.js",
+  "types": "dist/editor-social.cjs.d.ts",
   "files": [
     "style",
     "dist"
