+++ @remirror/editor-social/src/__tests__/social.spec.tsx
