+++ @remirror/editor-social/src/social-editor-types.ts
