+++ @remirror/editor-social/src/social-editor-utils.ts
