+++ @remirror/i18n/src/en/messages.po
@@ -8,6 +8,6 @@ msgstr ""
 "Language: en\n"
 
 #. The alt text for the user mention avatar.
-#: @remirror/react-editor-social/src/messages.ts:2
+#: @remirror/editor-social/src/messages.ts:2
 msgid "user.mention.avatar.alt"
 msgstr "Avatar for { name }"
