+++ @remirror/preset-react/src/__tests__/react-preset.spec.ts
@@ -1,6 +1,6 @@
 import { isPresetValid } from '@remirror/test-fixtures';
 
-import { ReactPreset } from '../../dist/preset-react.cjs';
+import { ReactPreset } from '..';
 
 test('is react preset valid', () => {
   expect(isPresetValid(ReactPreset, {}));
