+++ @remirror/showcase/jest.config.js
