+++ @remirror/showcase/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@remirror/react-showcase",
-  "version": "0.0.0",
+  "name": "@remirror/showcase",
+  "version": "0.13.1",
   "description": "A showcase of editors built with remirror and react.",
   "keywords": [
     "remirror",
@@ -8,16 +8,16 @@
     "text editor",
     "wysiwyg"
   ],
-  "homepage": "https://github.com/remirror/remirror/tree/next/@remirror/react-showcase",
-  "repository": "https://github.com/remirror/remirror/tree/master/@remirror/react-showcase",
+  "homepage": "https://github.com/remirror/remirror/tree/next/@remirror/showcase",
+  "repository": "https://github.com/remirror/remirror/tree/master/@remirror/showcase",
   "license": "MIT",
   "contributors": [
     "Ifiok Jr. <ifiokotung@gmail.com>"
   ],
   "sideEffects": false,
-  "main": "dist/react-showcase.cjs.js",
-  "module": "dist/react-showcase.esm.js",
-  "types": "dist/react-showcase.cjs.d.ts",
+  "main": "dist/showcase.cjs.js",
+  "module": "dist/showcase.esm.js",
+  "types": "dist/showcase.cjs.d.ts",
   "files": [
     "dist"
   ],
@@ -27,10 +27,10 @@
     "@remirror/core-extensions": "^0.13.1",
     "@remirror/dev": "^0.13.2",
     "@remirror/editor-markdown": "^0.13.1",
+    "@remirror/editor-social": "^0.13.1",
     "@remirror/extension-code-block": "^0.12.0",
     "@remirror/extension-epic-mode": "^0.11.0",
     "@remirror/react": "^0.13.1",
-    "@remirror/react-editor-social": "^0.0.0",
     "@remirror/ui-buttons": "^0.7.7",
     "@types/match-sorter": "^4.0.0",
     "@types/prettier": "^2.0.1",
