+++ @remirror/showcase/readme.md
@@ -1,4 +1,4 @@
-# @remirror/react-showcase
+# @remirror/showcase
 
 > A showcase of editors built with remirror and react.
 
@@ -6,25 +6,25 @@
 [![Bundled size][size-badge]][size] [![Typed Codebase][typescript]](./src/index.ts)
 ![MIT License][license]
 
-[version]: https://flat.badgen.net/npm/v/@remirror/react-showcase
-[npm]: https://npmjs.com/package/@remirror/react-showcase
+[version]: https://flat.badgen.net/npm/v/@remirror/showcase
+[npm]: https://npmjs.com/package/@remirror/showcase
 [license]: https://flat.badgen.net/badge/license/MIT/purple
-[size]: https://bundlephobia.com/result?p=@remirror/react-showcase
-[size-badge]: https://flat.badgen.net/bundlephobia/minzip/@remirror/react-showcase
+[size]: https://bundlephobia.com/result?p=@remirror/showcase
+[size-badge]: https://flat.badgen.net/bundlephobia/minzip/@remirror/showcase
 [typescript]: https://flat.badgen.net/badge/icon/TypeScript?icon=typescript&label
-[downloads-badge]: https://badgen.net/npm/dw/@remirror/react-showcase/red?icon=npm
+[downloads-badge]: https://badgen.net/npm/dw/@remirror/showcase/red?icon=npm
 
 ## Installation
 
 ```bash
 # yarn
-yarn add @remirror/react-showcase @remirror/pm
+yarn add @remirror/showcase @remirror/pm
 
 # pnpm
-pnpm add @remirror/react-showcase @remirror/pm
+pnpm add @remirror/showcase @remirror/pm
 
 # npm
-npm install @remirror/react-showcase @remirror/pm
+npm install @remirror/showcase @remirror/pm
 ```
 
 ## Credits
