+++ @remirror/styles/all.css
@@ -47,65 +47,7 @@
 }
 
 /**
- * Styles extracted from: @remirror/extension-blockquote/src/styles.ts
- */
-.remirror-editor blockquote {
-  border-left: 2px solid #ddd;
-  margin-left: 0;
-  margin-right: 0;
-  padding-left: 10px;
-  font-style: italic;
-}
-.remirror-editor blockquote p {
-  color: #888;
-}
-
-/**
- * Styles extracted from: @remirror/extension-gap-cursor/src/styles.ts
- */
-.remirror-editor .ProseMirror-gapcursor {
-  display: none;
-  pointer-events: none;
-  position: absolute;
-}
-.remirror-editor .ProseMirror-gapcursor:after {
-  content: '';
-  display: block;
-  position: absolute;
-  top: -2px;
-  width: 20px;
-  border-top: 1px solid black;
-  -webkit-animation: ProseMirror-cursor-blink-remirror-editor 1.1s steps(2, start) infinite;
-  animation: ProseMirror-cursor-blink-remirror-editor 1.1s steps(2, start) infinite;
-}
-@-webkit-keyframes ProseMirror-cursor-blink-remirror-editor {
-  to {
-    visibility: hidden;
-  }
-}
-@keyframes ProseMirror-cursor-blink-remirror-editor {
-  to {
-    visibility: hidden;
-  }
-}
-.remirror-editor .ProseMirror-focused .ProseMirror-gapcursor {
-  display: block;
-}
-
-/**
- * Styles extracted from: @remirror/extension-placeholder/src/placeholder-extension.ts
- */
-.remirror-is-empty:first-of-type::before {
-  position: absolute;
-  color: #aaa;
-  pointer-events: none;
-  height: 0;
-  font-style: italic;
-  content: attr(data-placeholder);
-}
-
-/**
- * Styles extracted from: @remirror/react-editor-social/src/social-editor-styles.ts
+ * Styles extracted from: @remirror/editor-social/src/social-editor-styles.ts
  */
 .remirror-emoji-suggestions-item {
   padding: 8px;
@@ -216,7 +158,7 @@
 }
 
 /**
- * Styles extracted from: @remirror/react-editor-social/src/components/social-editor-character-count.tsx
+ * Styles extracted from: @remirror/editor-social/src/components/social-editor-character-count.tsx
  */
 .remirror-character-count-wrapper {
   position: absolute;
@@ -252,7 +194,7 @@
 }
 
 /**
- * Styles extracted from: @remirror/react-editor-social/src/components/social-editor-emoji.tsx
+ * Styles extracted from: @remirror/editor-social/src/components/social-editor-emoji.tsx
  */
 .remirror-emoji-suggestions-item-name {
   color: rgb(121, 129, 134);
@@ -263,7 +205,7 @@
 }
 
 /**
- * Styles extracted from: @remirror/react-editor-social/src/components/social-editor.tsx
+ * Styles extracted from: @remirror/editor-social/src/components/social-editor.tsx
  */
 .remirror-social-editor .ProseMirror {
   -webkit-box-flex: 1;
@@ -321,4 +263,62 @@
 .remirror-social-editor-wrapper {
   position: relative;
   height: 100%;
+}
+
+/**
+ * Styles extracted from: @remirror/extension-blockquote/src/styles.ts
+ */
+.remirror-editor blockquote {
+  border-left: 2px solid #ddd;
+  margin-left: 0;
+  margin-right: 0;
+  padding-left: 10px;
+  font-style: italic;
+}
+.remirror-editor blockquote p {
+  color: #888;
+}
+
+/**
+ * Styles extracted from: @remirror/extension-gap-cursor/src/styles.ts
+ */
+.remirror-editor .ProseMirror-gapcursor {
+  display: none;
+  pointer-events: none;
+  position: absolute;
+}
+.remirror-editor .ProseMirror-gapcursor:after {
+  content: '';
+  display: block;
+  position: absolute;
+  top: -2px;
+  width: 20px;
+  border-top: 1px solid black;
+  -webkit-animation: ProseMirror-cursor-blink-remirror-editor 1.1s steps(2, start) infinite;
+  animation: ProseMirror-cursor-blink-remirror-editor 1.1s steps(2, start) infinite;
+}
+@-webkit-keyframes ProseMirror-cursor-blink-remirror-editor {
+  to {
+    visibility: hidden;
+  }
+}
+@keyframes ProseMirror-cursor-blink-remirror-editor {
+  to {
+    visibility: hidden;
+  }
+}
+.remirror-editor .ProseMirror-focused .ProseMirror-gapcursor {
+  display: block;
+}
+
+/**
+ * Styles extracted from: @remirror/extension-placeholder/src/placeholder-extension.ts
+ */
+.remirror-is-empty:first-of-type::before {
+  position: absolute;
+  color: #aaa;
+  pointer-events: none;
+  height: 0;
+  font-style: italic;
+  content: attr(data-placeholder);
 }
