+++ @remirror/styles/editor-social.css
@@ -1,5 +1,5 @@
 /**
- * Styles extracted from: @remirror/react-editor-social/src/social-editor-styles.ts
+ * Styles extracted from: @remirror/editor-social/src/social-editor-styles.ts
  */
 .remirror-emoji-suggestions-item {
   padding: 8px;
@@ -110,7 +110,7 @@
 }
 
 /**
- * Styles extracted from: @remirror/react-editor-social/src/components/social-editor-character-count.tsx
+ * Styles extracted from: @remirror/editor-social/src/components/social-editor-character-count.tsx
  */
 .remirror-character-count-wrapper {
   position: absolute;
@@ -146,7 +146,7 @@
 }
 
 /**
- * Styles extracted from: @remirror/react-editor-social/src/components/social-editor-emoji.tsx
+ * Styles extracted from: @remirror/editor-social/src/components/social-editor-emoji.tsx
  */
 .remirror-emoji-suggestions-item-name {
   color: rgb(121, 129, 134);
@@ -157,7 +157,7 @@
 }
 
 /**
- * Styles extracted from: @remirror/react-editor-social/src/components/social-editor.tsx
+ * Styles extracted from: @remirror/editor-social/src/components/social-editor.tsx
  */
 .remirror-social-editor .ProseMirror {
   -webkit-box-flex: 1;
