--- deprecated/editor-social/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@remirror/editor-social",
-  "version": "0.13.1",
-  "description": "DEPRECATED - A configurable social editor.",
-  "homepage": "https://github.com/remirror/remirror/tree/next/@remirror/react-editor-social",
-  "repository": "https://github.com/remirror/remirror/tree/master/deprecated/editor-social",
-  "license": "MIT",
-  "contributors": [
-    "Ifiok Jr. <ifiokotung@gmail.com>"
-  ],
-  "publishConfig": {
-    "access": "public"
-  }
-}
