--- deprecated/editor-social/readme.md
@@ -1,18 +1,0 @@
-# @remirror/editor-social
-
-This package is **deprecated** and has been moved to the
-[`@remirror/react-editor-social`](https://github.com/remirror/remirror/tree/next/@remirror/react-editor-social)
-package.
-
-It is now accessible with the following installation.
-
-```bash
-# Yarn
-yarn add @remirror/react-editor-social
-
-# PNPM
-pnpm add @remirror/react-editor-social
-
-# npm
-npm install @remirror/react-editor-social
-```
