--- deprecated/showcase/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@remirror/showcase",
-  "version": "0.13.1",
-  "description": "A showcase of editors built with remirror and react.",
-  "keywords": [
-    "remirror",
-    "react",
-    "text editor",
-    "wysiwyg"
-  ],
-  "homepage": "https://github.com/remirror/remirror/tree/next/@remirror/react-showcase",
-  "repository": "https://github.com/remirror/remirror/tree/master/deprecated/showcase",
-  "license": "MIT",
-  "contributors": [
-    "Ifiok Jr. <ifiokotung@gmail.com>"
-  ],
-  "publishConfig": {
-    "access": "public"
-  }
-}
