--- deprecated/showcase/readme.md
@@ -1,17 +1,0 @@
-# @remirror/showcase
-
-This package is **deprecated** and it's functionality is included by default with the
-`@remirror/react-showcase` package.
-
-It is now accessible with the following installation.
-
-```bash
-# Yarn
-yarn add @remirror/react-showcase
-
-# PNPM
-pnpm add @remirror/react-showcase
-
-# npm
-npm install @remirror/react-showcase
-```
