+++ docs/showcase/social.mdx
@@ -2,11 +2,11 @@
 title: Social Editor Showcase
 ---
 
-import { ExampleSocialEditor } from '@remirror/react-showcase';
+import { ExampleSocialEditor } from '@remirror/showcase';
 
 # Social Editor
 
-Below is an example of a social editor using the `@remirror/react-editor-social` package.
+Below is an example of a social editor using the `@remirror/editor-social` package.
 
 To test it out try adding an **`@`** mention or a **`#`** mention. Also typing **`:`** will activate
 the emoji search extension.
