+++ examples/with-next/src/pages/editor/social/index.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 
-import { ExampleSocialEditor } from '@remirror/react-showcase';
+import { ExampleSocialEditor } from '@remirror/showcase';
 
 const SocialEditor: FC = () => <ExampleSocialEditor suppressHydrationWarning={true} />;
 SocialEditor.displayName = 'SocialEditor';
